all: Decouple pkg/band tests
