+++ pkg/band/band.go
@@ -140,7 +140,7 @@ func (b Band) FindUplinkDataRate(dr *ttnpb.DataRate) (ttnpb.DataRateIndex, DataR
 		// NOTE(2): A more robust solution could be to record a list of uplink-only data rates
 		// per band and only match those here, however it is more complex and is not necessary.
 		bDR, ok := b.DataRates[i]
-		if ok && compareDatarates(bDR.Rate, dr, b.StrictCodingRate) {
+		if ok && compareDataRates(bDR.Rate, dr, b.StrictCodingRate) {
 			return i, bDR, true
 		}
 	}
@@ -155,7 +155,7 @@ func (b Band) FindDownlinkDataRate(dr *ttnpb.DataRate) (ttnpb.DataRateIndex, Dat
 	// NOTE: See notes in FindUplinkDataRate explaining the order of scanning data rates.
 	for i := ttnpb.DataRateIndex_DATA_RATE_15; i >= ttnpb.DataRateIndex_DATA_RATE_0; i-- {
 		bDR, ok := b.DataRates[i]
-		if ok && compareDatarates(bDR.Rate, dr, b.StrictCodingRate) {
+		if ok && compareDataRates(bDR.Rate, dr, b.StrictCodingRate) {
 			return i, bDR, true
 		}
 	}
@@ -164,10 +164,10 @@ func (b Band) FindDownlinkDataRate(dr *ttnpb.DataRate) (ttnpb.DataRateIndex, Dat
 
 func boolPtr(v bool) *bool { return &v }
 
-// compareDatarates checks if two given *ttnpb.DataRate's are equal. If the strict
+// compareDataRates checks if two given *ttnpb.DataRate's are equal. If the strict
 // bit is not set the *ttnpb.DataRate_Lora check will skip the `CodingRate`.
 // Make sure to update this check when altering any of the `DataRate` types.
-func compareDatarates(a, b *ttnpb.DataRate, strict bool) bool {
+func compareDataRates(a, b *ttnpb.DataRate, strict bool) bool {
 	switch modA := a.Modulation.(type) {
 	case *ttnpb.DataRate_Lora:
 		switch modB := b.Modulation.(type) {
